+++ ripflow/core/supervisor.py
@@ -1,7 +1,9 @@
-from .utils import Child
+from ripflow.core.utils import Child
+
+
 import time
-from threading import Thread
 from datetime import datetime, timedelta
+from threading import Thread
 
 
 class RestartPolicy:
